requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

torch>=1.12.0

wandb>=0.13.0
tqdm
pyarrow
pytest
PyYAML